APP_Django/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
Automat==24.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
constantly==23.10.4
coverage==7.6.9
Django==5.1.4
django-cors-headers==4.7.0
django-prometheus==2.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docstring-to-markdown==0.15
drf-yasg==1.21.8
greenlet==3.1.1
gunicorn==23.0.0
guppy3==3.1.4.post1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jedi==0.19.2
line_profiler==4.2.0
packaging==24.2
parso==0.8.4
pluggy==1.5.0
prometheus_client==0.21.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.20.0
SQLAlchemy==2.0.38
sqlparse==0.5.2
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
zope.interface==7.2
